buffer size is int

+++ src/main/scala/rxscalajs/subjects/ReplaySubject.scala
@@ -9,8 +9,8 @@ class ReplaySubject[T] protected(inner: ReplaySubjectFacade[T]) extends Subject[
 
 object ReplaySubject {
   def apply[T](): ReplaySubject[T] = new ReplaySubject(new ReplaySubjectFacade())
-  def withSize[T](bufferSize: Int): ReplaySubject[T] = new ReplaySubject(new ReplaySubjectFacade(bufferSize.toDouble))
+  def withSize[T](bufferSize: Int): ReplaySubject[T] = new ReplaySubject(new ReplaySubjectFacade(bufferSize))
   def withTime[T](time: Int, scheduler: Scheduler): ReplaySubject[T] = new ReplaySubject(new ReplaySubjectFacade(windowTime = time.toDouble, scheduler = scheduler))
-  def withTimeAndSize[T](time: Int, size: Int, scheduler: Scheduler): ReplaySubject[T] = new ReplaySubject(new ReplaySubjectFacade(size.toDouble,time.toDouble,scheduler))
+  def withTimeAndSize[T](time: Int, size: Int, scheduler: Scheduler): ReplaySubject[T] = new ReplaySubject(new ReplaySubjectFacade(size, time.toDouble,scheduler))
 }
